Named.Internal

Documentation

newtype NamedF f (a :: Type) (name :: Symbol)

Assign a name to a value of type a wrapped in f.

#verbose True :: NamedF Identity Bool "verbose"

Constructors

ArgF (f a)
Match on an F-argument without specifying its name. See also: argF.

pattern Arg :: a -> name :! a

Match on an argument without specifying its name. See also: arg.

type (:!) name a = NamedF Identity a name

Infix notation for the type of a named parameter.

type (:?) name a = NamedF Maybe a name

Infix notation for the type of an optional named parameter.

class InjValue f where

Methods

injValue :: a -> f a

newtype Param p

Constructors

Param p

param :: Name name -> a -> Param (name :! a)

Explicitly build a function parameter:

fn ! param #param_name value

This is equivalent to the implicit notation:

fn ! #param_name value

paramF :: Name name -> f a -> Param (NamedF f a name)

Explicitly build a function parameter inside an arity wrapper:

fn ! paramF #param_name (Identity value)
fn ! paramF #param_name (Just value)
fn ! paramF #param_name Nothing

This has no equivalent implicit notation.

(!) :: forall p fn fn'. WithParam p fn fn' => fn -> Param p -> fn' infixl 9

Supply a parameter to a function:

function ! #param_name value

function ! #x 7 ! #y 42 ! defaults

This is an infix version of with.

class WithParam p fn fn' | p fn -> fn' where

Supply a parameter p to a function fn, resulting in fn'.

For example, when we pass a single named parameter, we get a function without this parameter:

WithParam
                 ("x" :! Char)       -- p
  ("b" :! Bool -> "x" :! Char -> r)  -- fn
  ("b" :! Bool                -> r)  -- fn'

In case the parameter cannot be supplied, this constraint will become a type error.

Methods

with :: Param p -> fn -> fn'

Supply a parameter to a function:

 with (#param_name value) function
 

 with defaults function
 

This is a prefix version of the (!) operator.

data Defaults

Constructors

Defaults

defaults :: Param Defaults

Passing defaults to a function fills all unspecified optional parameters with Nothing:

fn            :: "b" :! Bool -> "x" :? Char -> Int -> IO ()
fn ! defaults :: "b" :! Bool ->                Int -> IO ()

data Name (name :: Symbol)

A proxy for a name, intended for use with -XOverloadedLabels:

#verbose :: Name "verbose"

Constructors

Name

arg :: Name name -> (name :! a) -> a

arg unwraps a named parameter with the specified name. One way to use it is to match on arguments with -XViewPatterns:

fn (arg #t -> t) (arg #f -> f) = ...

This way, the names of parameters can be inferred from the patterns: no type signature for fn is required. In case a type signature for fn is provided, the parameters must come in the same order:

fn :: "t" :! Integer -> "f" :! Integer -> ...
fn (arg #t -> t) (arg #f -> f) = ... -- ok
fn (arg #f -> f) (arg #t -> t) = ... -- does not typecheck

argF :: Name name -> NamedF f a name -> f a

argF is similar to arg: it unwraps a named parameter with the specified name. The difference is that the result of argF is inside an arity wrapper, which is Identity for normal parameters and Maybe for optional parameters.

argDef :: Name name -> a -> (name :? a) -> a

A variation of arg for optional arguments. Requires a default value to handle the case when the optional argument was omitted:

fn (argDef #answer 42 -> ans) = ...

In case you want to get a value wrapped in Maybe instead, use argF or ArgF.

data DApply

data DFill

data DPass

type family Decide (p :: Type) (fn :: Type) :: [Type] where ...

Equations

Decide (NamedF f' a' name) (NamedF f a name -> r) = '[DApply]

Decide Defaults (NamedF Maybe a name -> r) = DFill ': Decide Defaults r

Decide p (x -> r) = DPass ': Decide p r

Decide (NamedF f' a' name) t = TypeError ((Text "Named parameter '" :<>: Text name) :<>: Text "' was supplied, but not expected")

Decide Defaults t = '[]

class WithParam' (ds :: [Type]) p fn fn' | ds p fn -> fn' where

Methods

withParam :: p -> fn -> fn'